Run reaper less often, import more often
After testing the reaper further, I saw that it can take a very long
time to run through every record. For that reason, running it once
a month should be sufficient.

The import, which also acts as a monitor, can run daily. It's pretty
slow too, but it's still much faster than the reaper.