Add Dockerfile.webapp
Adding this Dockerfile allows the deprecated nypl/circ-deploy and the
new nypl/circ-webapp scripts to coexist for a time. Dockerfile.webapp &
Dockerfile.deploy need to have the same content.

This change is being made because the container name nypl/circ-deploy is
confusing, especially alongside nypl/circ-scripts.

Resolves NYPL-Simplified/circulation-docker#65, where more thoughts on
this issue can be found.